Add Graphing module
